Fix lint by moving Gemini OAuth recovery helpers to extension
Move OAuth discovery helpers out of the main probe struct and split
recovery API tests into their own file to satisfy type_body_length.
